Cache Busting Fix for Safari
Discovered an issue for Safari users where they could initially view my PDF contents, but the content would appear blank if they left and then returned. This added JavaScript appends a dynamic value based on the datetime, so each time the modal opens, it loads a new document. Unfortunately, this causes the user to re-download the content each time, but it is ultimately better than being unable to access the content at all due to their browser.